% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDensity}
\alias{plotDensity}
\title{Density of each ADT, raw counts overlapped with decontaminated counts}
\usage{
plotDensity(counts, decontaminated_counts, features)
}
\arguments{
\item{counts}{raw count matrix}

\item{decontaminated_counts}{decontaminated count matrix}

\item{features}{names of ADT to plot}
}
\value{
pdf in working directory
}
\description{
Density of each ADT, raw counts overlapped with decontaminated counts
}
