% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{process_stan_vb}
\alias{process_stan_vb}
\title{Process Stan output.}
\usage{
process_stan_vb(stan_vb_output)
}
\arguments{
\item{stan_vb_output}{}
}
\value{
Decomposed counts based on Stan estimate.
}
\description{
Process Stan output.
}
