% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peanutXdecontaminate.R
\name{peanutXdecontaminate}
\alias{peanutXdecontaminate}
\alias{peanutXdecontaminate,SingleCellExperiment-method}
\alias{peanutXdecontaminate,ANY-method}
\title{Decontaminate using peanutX}
\usage{
peanutXdecontaminate(object, cell_type, delta_sd, background_sd, ...)

\S4method{peanutXdecontaminate}{SingleCellExperiment}(object, cell_type, delta_sd, background_sd, ...)

\S4method{peanutXdecontaminate}{ANY}(object, cell_type, delta_sd, background_sd, ...)
}
\arguments{
\item{object}{Data matrix NxM (feature x droplet).}

\item{cell_type}{1xM vector of cell_type. 1-based.}

\item{delta_sd}{Prior for delta. Set to 2e-5 for now.}

\item{background_sd}{Prior for background. Set to 2e-6 for now.}

\item{...}{Additional arguments for generics.}
}
\value{
A list of decontaminated counts, and estimated parameters.
}
\description{
Decontaminate using peanutX
}
