% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peanutXdecontaminate.R
\name{peanutXdecontaminate}
\alias{peanutXdecontaminate}
\title{peanutX decontamination}
\usage{
peanutXdecontaminate(counts, cell_type, delta_sd, ground_sd)
}
\arguments{
\item{counts}{NxM sparse matrix, N ADTs and M droplets.}

\item{cell_type}{vector length M. 1 based.}

\item{ground_sd}{}
}
\value{
A list of decontamination estimation.
}
\description{
peanutX decontamination
}
