% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peanutXdecontaminate.R
\name{peanutXdecontaminate}
\alias{peanutXdecontaminate}
\title{Decontaminate using peanutX}
\usage{
peanutXdecontaminate(object, cell_type, delta_sd, background_sd, ...)
}
\arguments{
\item{object}{Data matrix NxM of single cell experiment class}

\item{cell_type}{1xM vector of cell_type. 1-based.}

\item{delta_sd}{Prior for delta. Set to 2e-5 for now.}

\item{background_sd}{Prior for background. Set to 2e-6 for now.}

\item{...}{}
}
\description{
Decontaminate using peanutX
}
